Create and use a shaded pass-authz-core library
Classes cannot be cast between class loaders.  In order to have one
ServletContext publish an AuthUserService that can be used by another
ServletContext, the AuthUserSevice needs to use a single class loader.  To
do this, pass-auth-core is now promoted to Tomcat's /lib directory, and
the scope of the dependency is changed to `provided` in the roles filter
and user service