Use submitterEmail() algorithm for user tokens
Previously, the user token encoded the submitter URI and a placeholder
URI.  Since emmber cannot deal with the placeholder, the model was
updated to add fields submitterName and submitterEmail, which
essentially serve as single-purpose placeholders.  Thus PR adappts the
user token algorithm to these changes.

Note:  The new mechanism is less general, and hard-wires substitution so
that it can _only_ apply to submitters (e.g. it would not be possible to
invite proxy users without a data model change).