Move token service out of the new user generator
Tokens are used for enacting permissions changes to submissions, and
have nothing to do with creating or updating users.  This conflation
lead to the incorrect solution of invalidating the cache in order to
enact a token.  Instead, token enaction should be entirely separate, and
protected by a cache so that it only happens once for a given token.